refactor back into helper, support SDL declaration.
