+++ packages/stitch/src/index.ts
@@ -1,4 +1,3 @@
 export { stitchSchemas } from './stitchSchemas';
 export { forwardArgsToSelectionSet } from './selectionSetArgs';
-
-export { ComputedFieldsPlugin } from './plugins/computedFieldsPlugin';
+export { isolateComputedMergeSchemas } from './isolateComputedMergeSchemas';
